requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
alembic>=1.12.0
jinja2>=3.1.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pandas>=2.0.0
email-validator>=2.0.0
httpx>=0.25.0
Pillow>=10.0.0
aiofiles>=23.0.0 